ZK_File_Vault/requirements.txt
fastapi
uvicorn
cryptography
argon2-cffi
requests